HiringTool/requirements.txt
streamlit>=1.39.0
anthropic>=0.39.0
pandas>=2.0.0
plotly>=5.18.0
PyPDF2>=3.0.0
python-docx>=1.0.0
